Renamed to folder OffcanvasAutoClose to MenuOffcanvas and NavItem to MenuItem
