+++ BlazorBootstrap53/Features/Home/MenuOffcanvas/BookSectionDynamicComponent.razor
@@ -8,15 +8,14 @@ else
 }
 
 
-
 @code {
 
-	[Parameter, EditorRequired] public Enums.NavItem? NavItem { get; set; }
+	[Parameter, EditorRequired] public Enums.MenuItem? MenuItem { get; set; }
 
 	protected Type? SelectedBookSectionCard;
 
 	protected override void OnParametersSet()
 	{
-		SelectedBookSectionCard = Type.GetType($"{DynamicComponentPaths.BookSectionCards}{NavItem!.Name}");
+		SelectedBookSectionCard = Type.GetType($"{DynamicComponentPaths.BookSectionCards}{MenuItem!.Name}");
 	}
 }
