+++ BlazorBootstrap53/Features/Home/MenuOffcanvas/BookSections/ChristianPassOver.razor
@@ -1,0 +1,7 @@
+﻿<h3>Christian Pass Over</h3>
+
+<p><b>ToDo</b>: Add content</p>
+
+@code {
+
+}
