+++ BlazorBootstrap53/Features/Home/MenuOffcanvas/BookSections/EarliestTempleAltar.razor
@@ -1,0 +1,7 @@
+﻿<h3>Earliest Temple Altar</h3>
+
+<p><b>ToDo</b>: Add content</p>
+
+@code {
+
+}
