+++ BlazorBootstrap53/Features/Home/MenuOffcanvas/BookSections/HebrewPassOverOrCrossOver.razor
@@ -1,0 +1,7 @@
+﻿<h3>Hebrew PassOver or CrossOver</h3>
+
+<p><b>ToDo</b>: Add content</p>
+
+@code {
+
+}
