+++ BlazorBootstrap53/Features/Home/MenuOffcanvas/BookSections/OriginOfTheRite.razor
@@ -1,0 +1,7 @@
+﻿<h3>Origin Of The Rite</h3>
+
+<p><b>ToDo</b>: Add content</p>
+
+@code {
+
+}
