+++ BlazorBootstrap53/Features/Home/MenuOffcanvas/BookSections/OutgrowthsAndPerversions.razor
@@ -1,0 +1,7 @@
+﻿<h3>Out growths and Perversions</h3>
+
+<p><b>ToDo</b>: Add content</p>
+
+@code {
+
+}
