+++ BlazorBootstrap53/Features/Home/MenuOffcanvas/BookSections/PrimativeFamilyAltar.razor
@@ -1,0 +1,7 @@
+﻿<h3>Primative Family Altar</h3>
+
+<p><b>ToDo</b>: Add content</p>
+
+@code {
+
+}
