+++ BlazorBootstrap53/Features/Home/MenuOffcanvas/BookSections/SacredBoundaryLine.razor
@@ -1,0 +1,7 @@
+﻿<h3>Sacred Boundary Line</h3>
+
+<p><b>ToDo</b>: Add content</p>
+
+@code {
+
+}
