+++ BlazorBootstrap53/Features/Home/MenuOffcanvas/Constants.cs
@@ -1,0 +1,7 @@
+﻿namespace BlazorBootstrap53.Features.Home.MenuOffcanvas;
+
+public static class DynamicComponentPaths
+{
+    public static string BookSectionCards = "BlazorBootstrap53.Features.Home.MenuOffcanvas.BookSections.";
+    //public static string SubSectionCards = "BlazorBootstrap53.Features.Home.MenuOffcanvas.BookSections.SubSections.";
+}
