--- BlazorBootstrap53/Features/Home/OffcanvasAutoClose/BookSections/ChristianPassOver.razor
@@ -1,5 +1,0 @@
-﻿<h3>ChristianPassOver</h3>
-
-@code {
-
-}
