--- BlazorBootstrap53/Features/Home/OffcanvasAutoClose/BookSections/EarliestTempleAltar.razor
@@ -1,5 +1,0 @@
-﻿<h3>EarliestTempleAltar</h3>
-
-@code {
-
-}
