--- BlazorBootstrap53/Features/Home/OffcanvasAutoClose/BookSections/HebrewPassOverOrCrossOver.razor
@@ -1,5 +1,0 @@
-﻿<h3>HebrewPassOverOrCrossOver</h3>
-
-@code {
-
-}
