--- BlazorBootstrap53/Features/Home/OffcanvasAutoClose/BookSections/OriginOfTheRite.razor
@@ -1,5 +1,0 @@
-﻿<h3>OriginOfTheRite</h3>
-
-@code {
-
-}
