--- BlazorBootstrap53/Features/Home/OffcanvasAutoClose/BookSections/OutgrowthsAndPerversions.razor
@@ -1,5 +1,0 @@
-﻿<h3>OutgrowthsAndPerversions</h3>
-
-@code {
-
-}
