--- BlazorBootstrap53/Features/Home/OffcanvasAutoClose/BookSections/PrimativeFamilyAltar.razor
@@ -1,5 +1,0 @@
-﻿<h3>PrimativeFamilyAltar</h3>
-
-@code {
-
-}
