--- BlazorBootstrap53/Features/Home/OffcanvasAutoClose/BookSections/SacredBoundaryLine.razor
@@ -1,5 +1,0 @@
-﻿<h3>SacredBoundaryLine</h3>
-
-@code {
-
-}
