--- BlazorBootstrap53/Features/Home/OffcanvasAutoClose/Constants.cs
@@ -1,7 +1,0 @@
-﻿namespace BlazorBootstrap53.Features.Home.OffcanvasAutoClose;
-
-public static class DynamicComponentPaths
-{
-    public static string BookSectionCards = "BlazorBootstrap53.Features.Home.OffcanvasAutoClose.BookSections.";
-    //public static string SubSectionCards = "BlazorBootstrap53.Features.Home.OffcanvasAutoClose.BookSections.SubSections.";
-}
