+++ BlazorBootstrap53/Shared/NavBar/HomeButton.razor
@@ -37,7 +37,7 @@
 	const string NavLink = "";  // " nav-link"
 	const string HomeIndex = "/";
 	const string Icon = "fas fa-chevron-right";  //fas fa-home
-	const string Title = "BS 5.3";
+	const string Title = "Home | Bootstrap 5.3";
 	const string TextColor = ""; // "text-danger";  
 																//@using BlazorBootstrap53.Links;
 	//href="@Home.Index"
